Added file_saver.html
You put a file in, it downloads to your device and it's fully offline. Made with ChatGPT.

What's the point? I had a problem with an offline device and the only way to fix my problem was to forcibly download the file. It wouldn't work by simply moving it to the download folder. So if you want to use it for something, or maybe you have the same issue go ahead.